refactor: move utils to transport/utils/, remove compat aliases, fix IFlowTransport imports
- Move flowRecoveryUtils.ts → utils/flowRecovery.ts per smncd suggestion
- Remove all backward-compatible aliases (no existing consumers)
- Fix broken imports in IFlowTransport.ts (was ./types/X, should be ./X)

All 78 transport tests pass.
